backend/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
aiofiles==24.1.0
    # via -r requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
alembic==1.16.4
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via aiohttp
bcrypt==4.3.0
    # via passlib
black==25.1.0
    # via -r requirements-dev.in
blis==1.3.0
    # via thinc
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   -r requirements.in
    #   black
    #   typer
    #   uvicorn
cloudpathlib==0.21.1
    # via weasel
confection==0.1.5
    # via
    #   thinc
    #   weasel
cryptography==45.0.6
    # via
    #   -r requirements.in
    #   python-jose
cymem==2.0.11
    # via
    #   preshed
    #   spacy
    #   thinc
dnspython==2.7.0
    # via pymongo
ecdsa==0.19.1
    # via python-jose
fastapi==0.116.1
    # via -r requirements.in
flake8==7.3.0
    # via -r requirements-dev.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
gputil==1.4.0
    # via -r requirements.in
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements-dev.in
    #   -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via spacy
joblib==1.5.1
    # via scikit-learn
langcodes==3.5.0
    # via spacy
language-data==1.3.0
    # via langcodes
lxml==6.0.0
    # via python-docx
mako==1.3.10
    # via alembic
marisa-trie==1.2.1
    # via language-data
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
murmurhash==1.0.13
    # via
    #   preshed
    #   spacy
    #   thinc
mypy==1.17.1
    # via -r requirements-dev.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
numpy==2.3.2
    # via
    #   -r requirements.in
    #   blis
    #   pandas
    #   scikit-learn
    #   scipy
    #   spacy
    #   thinc
packaging==25.0
    # via
    #   black
    #   pytesseract
    #   pytest
    #   spacy
    #   thinc
    #   weasel
pandas==2.3.1
    # via -r requirements.in
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pathspec==0.12.1
    # via
    #   black
    #   mypy
pdf2image==1.17.0
    # via -r requirements.in
pillow==11.3.0
    # via
    #   -r requirements.in
    #   pdf2image
    #   pytesseract
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via pytest
preshed==3.0.10
    # via
    #   spacy
    #   thinc
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycodestyle==2.14.0
    # via flake8
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   confection
    #   fastapi
    #   pydantic-settings
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt==2.10.1
    # via -r requirements.in
pymongo==4.14.0
    # via -r requirements.in
pypdf2==3.0.1
    # via -r requirements.in
pytesseract==0.3.13
    # via -r requirements.in
pytest==8.4.1
    # via
    #   -r requirements-dev.in
    #   pytest-asyncio
pytest-asyncio==1.1.0
    # via -r requirements-dev.in
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.2.0
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-jose[cryptography]==3.5.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   uvicorn
redis==6.4.0
    # via -r requirements.in
requests==2.32.4
    # via
    #   -r requirements.in
    #   spacy
    #   weasel
rich==14.1.0
    # via
    #   -r requirements.in
    #   typer
rsa==4.9.1
    # via python-jose
scikit-learn==1.7.1
    # via -r requirements.in
scipy==1.16.1
    # via scikit-learn
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
smart-open==7.3.0.post1
    # via weasel
sniffio==1.3.1
    # via anyio
spacy==3.8.7
    # via -r requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
sqlalchemy==2.0.42
    # via
    #   -r requirements.in
    #   alembic
srsly==2.5.1
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
starlette==0.47.2
    # via
    #   -r requirements.in
    #   fastapi
thinc==8.3.6
    # via spacy
threadpoolctl==3.6.0
    # via scikit-learn
tqdm==4.67.1
    # via spacy
typer==0.16.0
    # via
    #   spacy
    #   weasel
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   alembic
    #   anyio
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   -r requirements.in
    #   requests
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
watchfiles==1.1.0
    # via uvicorn
weasel==0.4.1
    # via spacy
websockets==15.0.1
    # via uvicorn
wrapt==1.17.2
    # via smart-open
yarl==1.20.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
